move tests to test_conversation_memory
